рефакторинг, начал избранное

+++ src/Components/Auth/http/authApi.js
@@ -16,7 +16,7 @@ export const registrationEmail = async (email, password, username) => {
   }
 };
 
-//Получить код для подтверждения регистрации
+//Получить код для подтверждения регистрации с емейл
 export const confirmRegistrationEmail = async (
   registrant_id,
   verification_code,
@@ -46,6 +46,39 @@ export const loginEmail = async (email, password) => {
     const res = await axios.post(`${BASE_URL}/auth/login/email`, {
       email,
       password,
+    });
+    localStorage.setItem("user_id", res.data.user_id);
+    localStorage.setItem("access_token", res.data.access_token);
+    localStorage.setItem("refresh_token", res.data.refresh_token);
+    return jwt_decode(res.data.access_token);
+  } catch (err) {
+    console.log(err);
+  }
+};
+
+//вход/регистрация с телефона
+export const registrationPhone = async (phoneNumber) => {
+  try {
+    const res = await axios.post(`${BASE_URL}/auth/login/phone/whatsapp/code`, {
+      phoneNumber,
+    });
+    return res;
+  } catch (err) {
+    console.log(err);
+  }
+};
+
+//Получить код для подтверждения регистрации с телефона
+export const confirmRegistrationPhone = async (
+  verification_code,
+  phone_number,
+  firebase_id = phone_number
+) => {
+  try {
+    const res = await axios.post(`${BASE_URL}/auth/login/phone/code/confirm`, {
+      verification_code,
+      phone_number,
+      firebase_id,
     });
     localStorage.setItem("user_id", res.data.user_id);
     localStorage.setItem("access_token", res.data.access_token);
